Add files via upload
I changed the constructor of Room to include lookText, a string.
I changed the createRooms method in Game to include lookText for each room. 
In the Game class, line 263 onward, I made it so that entering "look" returns the lookText for the room.
In the goRoom method, I fixed the coffee stuff. Now, you can't enter Tannen's room without having had coffee.
In addition, I added the Hot Pocket, which restores your health to 10 when you enter the diner, which is the room just before Tannen's room. This means you'll get healed right before the final battle, every time.
I also made a southward exit for the secret area, so that you can go straight to the hideout. You can also skip coffee this way, as going to the secret area sets the coffee boolean to true. You don't need coffee if you haven't had drinks.